move base-table to core folder
