+++ src/components/tableV2/core/headCell.js
@@ -8,7 +8,7 @@ import Box from "src/components/templates/box"
 import Tooltip from "src/components/drops/tooltip"
 import { Icon } from "src/components/icon"
 
-import Table from "../components/base-table"
+import Table from "./base-table"
 
 import { flexRender } from "@tanstack/react-table"
